connexion et protection admin faites

+++ app/src/Controller/AdminController.php
@@ -23,9 +23,19 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class AdminController extends AbstractController
 {
+    private UserRepository $uRep;
+
+    public function __construct(UserRepository $uRep) {
+        $this->uRep = $uRep;
+    }
     #[Route('/admin', name: 'app_admin')]
     public function index(): Response
     {
+        $user = $this->uRep->find($this->getUser());
+        if ($user==null|| !in_array('admin',$user->getRoles()) ){
+            return $this->redirectToRoute('app_accueil');
+        }
+
         return $this->render('admin/index.html.twig', [
 
         ]);
@@ -34,6 +44,11 @@ class AdminController extends AbstractController
     #[Route('/admin/user', name: 'app_admin_user')]
     public function user(UserRepository $uRep, Request $request, EntityManagerInterface $entityManager, UserPasswordHasherInterface $userPasswordHasher): Response
     {
+        $user = $this->uRep->find($this->getUser());
+        if ($user==null|| !in_array('admin',$user->getRoles()) ){
+            return $this->redirectToRoute('app_accueil');
+        }
+
         $user = New User();
         $addUserForm = $this->createForm(AdminAddUserType::class, $user);
         $addUserForm->handleRequest($request);
@@ -61,6 +76,11 @@ class AdminController extends AbstractController
     #[Route('/admin/user/{user_id}', name: 'app_admin_user_remove')]
     public function userDelete(int $user_id, UserRepository $uRep, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->uRep->find($this->getUser());
+        if ($user==null|| !in_array('admin',$user->getRoles()) ){
+            return $this->redirectToRoute('app_accueil');
+        }
+
         $entityManager->remove($uRep->find($user_id));
         $entityManager->flush();
         return $this->redirectToRoute('app_admin_user');
@@ -71,6 +91,11 @@ class AdminController extends AbstractController
     #[Route('/admin/article', name: 'app_admin_article')]
     public function article(ArticleRepository $aRep, Request $request, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->uRep->find($this->getUser());
+        if ($user==null|| !in_array('admin',$user->getRoles()) ){
+            return $this->redirectToRoute('app_accueil');
+        }
+
         $addProduitForm = $this->createForm(AdminAddArticleType::class, new Article);
         $addProduitForm->handleRequest($request);
         if ($addProduitForm->isSubmitted() && $addProduitForm->isValid()) {
@@ -86,6 +111,11 @@ class AdminController extends AbstractController
     #[Route('/admin/article/{article_id}', name: 'app_admin_article_remove')]
     public function articleDelete(int $article_id, ArticleRepository $aRep, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->uRep->find($this->getUser());
+        if ($user==null|| !in_array('admin',$user->getRoles()) ){
+            return $this->redirectToRoute('app_accueil');
+        }
+
         $entityManager->remove($aRep->find($article_id));
         $entityManager->flush();
         return $this->redirectToRoute('app_admin_article');
@@ -96,6 +126,11 @@ class AdminController extends AbstractController
     #[Route('/admin/exemplaire', name: 'app_admin_exemplaire')]
     public function exemplaire(ExemplaireRepository $eRep, EntityManagerInterface $entityManager, Request $request): Response
     {
+        $user = $this->uRep->find($this->getUser());
+        if ($user==null|| !in_array('admin',$user->getRoles()) ){
+            return $this->redirectToRoute('app_accueil');
+        }
+
         $addExemplaireForm = $this->createForm(AdminAddExemplaireType::class, new Exemplaire);
         $addExemplaireForm->handleRequest($request);
         if ($addExemplaireForm->isSubmitted() && $addExemplaireForm->isValid()) {
@@ -111,6 +146,11 @@ class AdminController extends AbstractController
     #[Route('/admin/exemplaire/{exemplaire_id}', name: 'app_admin_exemplaire_remove')]
     public function exemplaireDelete(int $exemplaire_id, ExemplaireRepository $eRep, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->uRep->find($this->getUser());
+        if ($user==null|| !in_array('admin',$user->getRoles()) ){
+            return $this->redirectToRoute('app_accueil');
+        }
+        
         $entityManager->remove($eRep->find($exemplaire_id));
         $entityManager->flush();
         return $this->redirectToRoute('app_admin_exemplaire');
